Revert "Mitigate GUI DoS (part 1)"
First of all Ctrl+Alt+P is too common key combo (for example conflicts
with Firefox's "New Private Window"). But even if that wouldn't be true,
in the current shape it does more harm than good, because it is
not documented (hard to discover when actually needed), and also
pause system VMs which in case of sys-usb and USB keyboard is
fatal.

Lets rethink this first.

This reverts commit 9459331a3772decbaa3948cf6f45d98fd87ec800.
And adds a script to adjust user configuration.
QubesOS/qubes-issues#881
QubesOS/qubes-issues#4101
Fixes QubesOS/qubes-issues#4700