Use default title color, except for black
This is a partial revert of the change introduced in
d2e53b3d143d6fc7017fb946bfe84a6330246ec7.

The title colors look improve the text visibility for the default
theme (particularly inactive text); however, they make the sitation
worse if you want to use a different theme.

For now, override the title only in the black-on-black case to fix
the usability regression. In 4.1, I'll be trying to do something
more helpful to the user.

See discussion in QubesOS/qubes-issues#5800.