Update existing XFCE configurations
XFCE copies the default configuration into the user’s home directory,
and afterwards ignores changes to the defaults.  Fortunately,
‘update-xfce-config’ already takes care of updating existing
configurations.  Extend it to replace the default “xkill” keybinding
with “qvm-xkill”.

While I was working on this, I added locking to prevent concurrent
executions, and a Vim modeline so that the formatting is preserved.